requirements.txt
beautifulsoup4==4.12.3
feedparser==6.0.11
httpx==0.27.0
lxml==5.2.2
pandas==2.2.2
phonenumbers==8.13.42
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
pydantic==2.8.2
tqdm==4.66.4
urllib3==2.2.2
duckdb==1.0.0
apscheduler==3.10.4
streamlit==1.37.0
azure-identity==1.17.1
azure-ai-projects==1.0.0b7
azure-ai-agents==1.0.0b2
